#128 Fixed serialization issues with AutoValue classes and nullability of their fields
Some test methods failed due to incorrect serialization of the class Employee. Therefore the class JsonbConfigurator was introduced so Jsonb can access the fields of the auto-generated classes AutoValue_Employee and AutoValue_User. The fix with nullability of those fields, which was performed with the last commit, did not work so it was corrected now.