#151 Fixed bug that did not detect missing journey back
When the entry for a return journey is missing and no more journeys are booked after that in the current billing month, the algorithm did not detect the missing entry.
A journey gets invalidated now if the last project entry of this month did not end the journey.
One redundant test method in JourneyDirectionHandlerTest.java was removed and one more added.
Also test cases were implemented for the improved algorithm.