#173 changed month end report
The REST method WorkerResource.getMonthEndReports does not return 404 any more, but empty data when no project times were found.
This empty data is set in MonthlyReportServiceImpl.calcWarnings. The reason for this is that the interceptor in the frontend catches all 404 errors, but configuring exceptions for only the monthly report is unnecessarily complex. Also returning empty data instead of 404 is not a violation against the REST standard.